--- constants.py
@@ -1,8 +1,0 @@
-SCREEN_TITLE = "You better run..."
-INITIAL_SCREEN_WIDTH = 960
-INITIAL_SCREEN_HEIGHT = 482
-
-TILE_SPRITE_SCALING = 4
-PLAYER_SCALING = 1
-
-TILE_SIZE = 8 * TILE_SPRITE_SCALING
