+++ game/views/credits_view.py
