+++ game/views/instructions_view.py
