+++ game/views/menu_view.py
