--- item/key.py
@@ -1,5 +1,0 @@
-from model.item import Item
-
-
-class Key(Item):
-    ...
