--- model/interactable.py
@@ -1,11 +1,0 @@
-from abc import ABC, abstractmethod
-
-
-class Interactable(ABC):
-    @abstractmethod
-    def interact(self, player: "Player"):
-        """
-
-        :return:
-        """
-        raise NotImplementedError()
